+++ app/build.gradle
@@ -3,7 +3,7 @@ apply plugin: 'com.android.application'
 android {
     compileSdkVersion 28
     defaultConfig {
-        applicationId "io.display.androidsample"
+        applicationId "com.brandio.androidsample"
         minSdkVersion 14
         targetSdkVersion 28
         versionCode 1
