+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="io.display.androidsample">
+    package="com.brandio.androidsample">
 
     <application
         android:allowBackup="true"
@@ -9,16 +9,16 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".ShowListWithBannerActivity"></activity>
-        <activity android:name=".MainActivity">
+        <activity android:name="com.brandio.androidsample.ShowListWithBannerActivity"></activity>
+        <activity android:name="com.brandio.androidsample.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".BannerActivity" />
-        <activity android:name=".InterstitialActivity" />
+        <activity android:name="com.brandio.androidsample.BannerActivity" />
+        <activity android:name="com.brandio.androidsample.InterstitialActivity" />
     </application>
 
 </manifest>
