+++ app/src/main/java/com/brandio/androidsample/BannerListAdapter.java
@@ -1,4 +1,4 @@
-package io.display.androidsample;
+package com.brandio.androidsample;
 
 import android.content.Context;
 import android.support.annotation.NonNull;
@@ -12,6 +12,7 @@ import android.widget.TextView;
 
 import java.util.ArrayList;
 import java.util.List;
+
 
 import io.display.sdk.BannerPlacement;
 import io.display.sdk.Controller;
