+++ app/src/main/java/com/brandio/androidsample/InterstitialActivity.java
@@ -1,16 +1,12 @@
-package io.display.androidsample;
+package com.brandio.androidsample;
 
-import android.content.Intent;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
 import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.Toast;
 
-import java.util.ArrayList;
 
 import io.display.sdk.AdProvider;
 import io.display.sdk.AdRequest;
@@ -18,29 +14,28 @@ import io.display.sdk.Controller;
 import io.display.sdk.Placement;
 import io.display.sdk.ads.Ad;
 import io.display.sdk.exceptions.DioSdkException;
+import io.display.sdk.listeners.AdEventListener;
 import io.display.sdk.listeners.AdLoadListener;
 import io.display.sdk.listeners.AdRequestListener;
 
-public class BannerActivity extends AppCompatActivity {
+public class InterstitialActivity extends AppCompatActivity {
 
-    private static String TAG = "BannerActivity";
+    private static String TAG = "InterstitialActivity";
 
     private Button loadButton;
     private Button showButton;
 
     private String placementId;
-    private String requestId;
-    RecyclerView recyclerView;
+    private Ad loadedAd;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_banner);
-
+        setContentView(R.layout.activity_interstitial);
 
         placementId = getIntent().getStringExtra("placementId");
 
-        loadButton = findViewById(R.id.button_load_banner);
+        loadButton = findViewById(R.id.button_load_interstitial);
         loadButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
@@ -48,24 +43,13 @@ public class BannerActivity extends AppCompatActivity {
             }
         });
 
-        showButton = findViewById(R.id.button_show_banner);
+        showButton = findViewById(R.id.button_show_interstitial);
         showButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                Intent intent = new Intent(BannerActivity.this, ShowListWithBannerActivity.class);
-                intent.putExtra("placementId", placementId);
-                intent.putExtra("requestId", requestId);
-                startActivity(intent);
-                //showAd();
+                showAd();
             }
         });
-    }
-
-
-    @Override
-    protected void onResume() {
-        super.onResume();
-        loadButton.setEnabled(true);
     }
 
     private void loadAd() {
@@ -74,12 +58,13 @@ public class BannerActivity extends AppCompatActivity {
         Placement placement;
         try {
             placement = Controller.getInstance().getPlacement(placementId);
+            placement.getData();
         } catch (DioSdkException e) {
             Log.e(TAG, e.getLocalizedMessage());
             return;
         }
 
-        final AdRequest adRequest = placement.newAdRequest();
+        AdRequest adRequest = placement.newAdRequest();
         adRequest.setAdRequestListener(new AdRequestListener() {
             @Override
             public void onAdReceived(AdProvider adProvider) {
@@ -87,13 +72,13 @@ public class BannerActivity extends AppCompatActivity {
                 adProvider.setAdLoadListener(new AdLoadListener() {
                     @Override
                     public void onLoaded(Ad ad) {
-                        requestId = adRequest.getId();
+                        loadedAd = ad;
                         showButton.setEnabled(true);
                     }
 
                     @Override
                     public void onFailedToLoad() {
-                        Toast.makeText(BannerActivity.this, "Ad for placement " + placementId + " failed to load", Toast.LENGTH_LONG).show();
+                        Toast.makeText(InterstitialActivity.this, "Ad for placement " + placementId + " failed to load", Toast.LENGTH_LONG).show();
                     }
                 });
 
@@ -106,7 +91,7 @@ public class BannerActivity extends AppCompatActivity {
 
             @Override
             public void onNoAds() {
-                Toast.makeText(BannerActivity.this, "No Ads placement " + placementId, Toast.LENGTH_LONG).show();
+                Toast.makeText(InterstitialActivity.this, "No Ads placement " + placementId, Toast.LENGTH_LONG).show();
             }
         });
 
@@ -115,19 +100,34 @@ public class BannerActivity extends AppCompatActivity {
 
     private void showAd() {
         showButton.setEnabled(false);
-        setupRecyclerView();
-    }
 
-    private void setupRecyclerView() {
+        loadedAd.setEventListener(new AdEventListener() {
+            @Override
+            public void onShown(Ad ad) {
+                Log.e(TAG, "onShown");
+            }
 
-        recyclerView = findViewById(R.id.recycler_view_banner);
-        recyclerView.setLayoutManager(new LinearLayoutManager(this, LinearLayoutManager.VERTICAL, false));
+            @Override
+            public void onFailedToShow(Ad ad) {
+                Log.e(TAG, "onFailedToShow");
+            }
 
-        ArrayList<String> items = new ArrayList<>();
-        for (int i = 1; i <= 10; i++) {
-            items.add("Content item " + i);
-        }
+            @Override
+            public void onClicked(Ad ad) {
+                Log.e(TAG, "onClicked");
+            }
 
-        recyclerView.setAdapter(new BannerListAdapter(items, 2, placementId, requestId));
+            @Override
+            public void onClosed(Ad ad) {
+                Log.e(TAG, "onClosed");
+            }
+
+            @Override
+            public void onAdCompleted(Ad ad) {
+                Log.e(TAG, "onAdCompleted");
+            }
+        });
+
+        loadedAd.showAd(this);
     }
 }
