+++ app/src/main/java/com/brandio/androidsample/PlacementListAdapter.java
@@ -1,10 +1,11 @@
-package io.display.androidsample;
+package com.brandio.androidsample;
 
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
+
 
 public class PlacementListAdapter extends RecyclerView.Adapter<PlacementListAdapter.MyViewHolder> {
