+++ app/src/main/java/com/brandio/androidsample/ShowListWithBannerActivity.java
@@ -1,12 +1,13 @@
-package io.display.androidsample;
+package com.brandio.androidsample;
 
-import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 
 import java.util.ArrayList;
+
+
 
 public class ShowListWithBannerActivity extends AppCompatActivity {
