+++ app/src/main/res/layout/activity_banner.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".BannerActivity"
+    tools:context="com.brandio.androidsample.BannerActivity"
     android:orientation="vertical">
 
     <LinearLayout
