+++ app/src/main/res/layout/activity_interstitial.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".InterstitialActivity"
+    tools:context="com.brandio.androidsample.InterstitialActivity"
     android:orientation="horizontal">
 
     <Button
