+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context="com.brandio.androidsample.MainActivity">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/recycler_view_main"
