+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">Display.io sample app</string>
+    <string name="app_name">Brandio sample app</string>
 </resources>
